refactor: create-account and signup pages
