--- src/routes/create-account/+layout.svelte
@@ -1,3 +1,0 @@
-<div class="flex min-h-screen items-center justify-center">
-	<slot />
-</div>
