--- src/routes/create-account/signup/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import CreateAccount from './signup.svelte';
-</script>
-
-<div class="md:w-1/2 xl:w-1/3">
-	<CreateAccount />
-</div>
